pyproject.toml
[tool.poetry]
name = "miniircd-peerchat"
version = "1.3.0"
description = "Re-implementation of a GameSpy service that allows simple and easy, game-to-game chat."
authors = []
license = "GPL2"
readme = "README.md"
packages = [{include = "source"}]

[tool.poetry.dependencies]
google-re2 = "^1.0"
loguru = "^0.7.0"
python = "^3.7"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.264"
pyright = "^1.1.305"

[tool.poetry.scripts]
miniircd = "source.miniircd:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
